[android] moving related dialog inside the same file as the preference
