+++ android/res/layout/dialog_preference_number_picker2.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-/* Copyright (C) 2009-2011 Mike Novak <michael.novakjr@gmail.com>
+/* Copyright (C) 2017, FrostWire(R). All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
